feat: including combobox with graduation courses and admission year
