+++ components/datasets/collegeData.tsx
@@ -1,0 +1,485 @@
+export const collegeData = [
+    {
+      index: "1",
+      name: "UFGD",
+      label: "Fundação Universidade Federal da Grande Dourados",
+    },
+    {
+      index: "2",
+      name: "UFCSPA",
+      label: "Fundação Universidade Federal de Ciências da Saúde de Porto Alegre",
+    },
+    {
+      index: "3",
+      name: "UNIR",
+      label: "Fundação Universidade Federal de Rondônia",
+    },
+    { index: "4", name: "UFABC", label: "Fundação Universidade Federal do Abc" },
+    {
+      index: "5",
+      name: "UNIPAMPA",
+      label: "Fundação Universidade Federal do Pampa - Unipampa",
+    },
+    {
+      index: "6",
+      name: "UFT",
+      label: "Fundação Universidade Federal do Tocantins",
+    },
+    {
+      index: "7",
+      name: "UNIVASF",
+      label: "Fundação Universidade Federal do Vale do São Francisco",
+    },
+    {
+      index: "8",
+      name: "UDESC",
+      label: "Fundação Universidade do Estado de Santa Catarina",
+    },
+    {
+      index: "9",
+      name: "PUC-Campinas",
+      label: "Pontifícia Universidade Católica de Campinas",
+    },
+    {
+      index: "10",
+      name: "PUC Goiás",
+      label: "Pontifícia Universidade Católica de Goiás",
+    },
+    {
+      index: "11",
+      name: "PUC Minas",
+      label: "Pontifícia Universidade Católica de Minas Gerais",
+    },
+    {
+      index: "12",
+      name: "PUCSP",
+      label: "Pontifícia Universidade Católica de São Paulo",
+    },
+    {
+      index: "13",
+      name: "PUCPR",
+      label: "Pontifícia Universidade Católica do Paraná",
+    },
+    {
+      index: "14",
+      name: "PUCRS",
+      label: "Pontifícia Universidade Católica do Rio Grande do Sul",
+    },
+    {
+      index: "15",
+      name: "PUC Rio",
+      label: "Pontifícia Universidade Católica do Rio de Janeiro",
+    },
+    {
+      index: "16",
+      name: "UNIARP",
+      label: "Universidade Alto Vale do Rio do Peixe",
+    },
+    { index: "17", name: "UNIDERP", label: "Universidade Anhanguera" },
+    {
+      index: "18",
+      name: "UNIAN SP",
+      label: "Universidade Anhanguera de São Paulo",
+    },
+    { index: "19", name: "ANHEMBI", label: "Universidade Anhembi Morumbi" },
+    { index: "20", name: "UNIBR", label: "Universidade Brasil" },
+    { index: "21", name: "UCB RJ", label: "Universidade Castelo Branco" },
+    { index: "22", name: "UCDB", label: "Universidade Católica Dom Bosco" },
+    { index: "23", name: "UCB DF", label: "Universidade Católica de Brasília" },
+    { index: "24", name: "UCPel", label: "Universidade Católica de Pelotas" },
+    { index: "25", name: "UNICAP", label: "Universidade Católica de Pernambuco" },
+    { index: "26", name: "UCP RJ", label: "Universidade Católica de Petrópolis" },
+    { index: "27", name: "UNISANTOS", label: "Universidade Católica de Santos" },
+    { index: "28", name: "UCSAL", label: "Universidade Católica do Salvador" },
+    { index: "29", name: "UNICID", label: "Universidade Cidade de São Paulo" },
+    {
+      index: "30",
+      name: "UNOCHAPECÓ",
+      label: "Universidade Comunitária da Região de Chapecó",
+    },
+    { index: "31", name: "UNICSUL", label: "Universidade Cruzeiro do Sul" },
+    { index: "32", name: "UCAM", label: "Universidade Cândido Mendes" },
+    {
+      index: "33",
+      name: "UNESP",
+      label: "Universidade Estadual Paulista Júlio de Mesquita Filho",
+    },
+    { index: "34", name: "UEPB", label: "Universidade Estadual da Paraíba" },
+    { index: "35", name: "UNEAL", label: "Universidade Estadual de Alagoas" },
+    { index: "36", name: "UNICAMP", label: "Universidade Estadual de Campinas" },
+    {
+      index: "37",
+      name: "UNCISAL",
+      label: "Universidade Estadual de Ciências da Saúde de Alagoas",
+    },
+    {
+      index: "38",
+      name: "UEFS",
+      label: "Universidade Estadual de Feira de Santana",
+    },
+    { index: "39", name: "UEG", label: "Universidade Estadual de Goiás" },
+    { index: "40", name: "UEL", label: "Universidade Estadual de Londrina" },
+    { index: "41", name: "UEM", label: "Universidade Estadual de Maringá" },
+    {
+      index: "42",
+      name: "UEMS",
+      label: "Universidade Estadual de Mato Grosso do Sul",
+    },
+    {
+      index: "43",
+      name: "UNIMONTES",
+      label: "Universidade Estadual de Montes Claros",
+    },
+    { index: "44", name: "UEPG", label: "Universidade Estadual de Ponta Grossa" },
+    { index: "45", name: "UERR", label: "Universidade Estadual de Roraima" },
+    { index: "46", name: "UESC", label: "Universidade Estadual de Santa Cruz" },
+    { index: "47", name: "UECE", label: "Universidade Estadual do Ceará" },
+    {
+      index: "48",
+      name: "UNICENTRO PR",
+      label: "Universidade Estadual do Centro Oeste",
+    },
+    { index: "49", name: "UEMA", label: "Universidade Estadual do Maranhão" },
+    {
+      index: "50",
+      name: "UENF",
+      label: "Universidade Estadual do Norte Fluminense Darcy Ribeiro",
+    },
+    {
+      index: "51",
+      name: "UENP",
+      label: "Universidade Estadual do Norte do Paraná",
+    },
+    {
+      index: "52",
+      name: "UNIOESTE",
+      label: "Universidade Estadual do Oeste do Paraná",
+    },
+    { index: "53", name: "UNESPAR", label: "Universidade Estadual do Paraná" },
+    { index: "54", name: "UESPI", label: "Universidade Estadual do Piauí" },
+    {
+      index: "55",
+      name: "UERGS",
+      label: "Universidade Estadual do Rio Grande do Sul",
+    },
+    {
+      index: "56",
+      name: "UESB",
+      label: "Universidade Estadual do Sudoeste da Bahia",
+    },
+    {
+      index: "57",
+      name: "UVA CE",
+      label: "Universidade Estadual do Vale do Acaraú",
+    },
+    { index: "58", name: "UNESA", label: "Universidade Estácio de Sá" },
+    { index: "59", name: "UFF", label: "Universidade Federal Fluminense" },
+    {
+      index: "60",
+      name: "UFRA",
+      label: "Universidade Federal Rural da Amazônia",
+    },
+    {
+      index: "61",
+      name: "UFRPE",
+      label: "Universidade Federal Rural de Pernambuco",
+    },
+    {
+      index: "62",
+      name: "UFRRJ",
+      label: "Universidade Federal Rural do Rio de Janeiro",
+    },
+    {
+      index: "63",
+      name: "UFERSA",
+      label: "Universidade Federal Rural do Semi-Árido",
+    },
+    { index: "64", name: "UFBA", label: "Universidade Federal da Bahia" },
+    { index: "65", name: "UFFS", label: "Universidade Federal da Fronteira Sul" },
+    {
+      index: "66",
+      name: "UNILA",
+      label: "Universidade Federal da Integração Latino-Americana",
+    },
+    { index: "67", name: "UFPB", label: "Universidade Federal da Paraíba" },
+    { index: "68", name: "UFAL", label: "Universidade Federal de Alagoas" },
+    { index: "69", name: "UNIFAL MG", label: "Universidade Federal de Alfenas" },
+    {
+      index: "70",
+      name: "UFCG",
+      label: "Universidade Federal de Campina Grande",
+    },
+    { index: "71", name: "UFG", label: "Universidade Federal de Goiás" },
+    { index: "72", name: "UNIFEI", label: "Universidade Federal de Itajubá" },
+    { index: "73", name: "UFJF", label: "Universidade Federal de Juiz de Fora" },
+    { index: "74", name: "UFLA", label: "Universidade Federal de Lavras" },
+    { index: "75", name: "UFMT", label: "Universidade Federal de Mato Grosso" },
+    {
+      index: "76",
+      name: "UFMS",
+      label: "Universidade Federal de Mato Grosso do Sul",
+    },
+    { index: "77", name: "UFMG", label: "Universidade Federal de Minas Gerais" },
+    { index: "78", name: "UFOP", label: "Universidade Federal de Ouro Preto" },
+    { index: "79", name: "UFPel", label: "Universidade Federal de Pelotas" },
+    { index: "80", name: "UFPE", label: "Universidade Federal de Pernambuco" },
+    { index: "81", name: "UFRR", label: "Universidade Federal de Roraima" },
+    {
+      index: "82",
+      name: "UFSC",
+      label: "Universidade Federal de Santa Catarina",
+    },
+    { index: "83", name: "UFSM", label: "Universidade Federal de Santa Maria" },
+    { index: "84", name: "UFS", label: "Universidade Federal de Sergipe" },
+    { index: "85", name: "UFSCAR", label: "Universidade Federal de São Carlos" },
+    {
+      index: "86",
+      name: "UFSJ",
+      label: "Universidade Federal de São João Del Rei",
+    },
+    { index: "87", name: "UNIFESP", label: "Universidade Federal de São Paulo" },
+    { index: "88", name: "UFU", label: "Universidade Federal de Uberlândia" },
+    { index: "89", name: "UFV", label: "Universidade Federal de Viçosa" },
+    { index: "90", name: "UFAC", label: "Universidade Federal do Acre" },
+    { index: "91", name: "UNIFAP", label: "Universidade Federal do Amapá" },
+    { index: "92", name: "UFAM", label: "Universidade Federal do Amazonas" },
+    { index: "93", name: "UFCA", label: "Universidade Federal do Cariri" },
+    { index: "94", name: "UFC", label: "Universidade Federal do Ceará" },
+    {
+      index: "95",
+      name: "UFES",
+      label: "Universidade Federal do Espírito Santo",
+    },
+    {
+      index: "96",
+      name: "UNIRIO",
+      label: "Universidade Federal do Estado do Rio de Janeiro",
+    },
+    { index: "97", name: "UFMA", label: "Universidade Federal do Maranhão" },
+    {
+      index: "98",
+      name: "UFOB",
+      label: "Universidade Federal do Oeste da Bahia",
+    },
+    {
+      index: "99",
+      name: "UFOPA",
+      label: "Universidade Federal do Oeste do Pará",
+    },
+    { index: "100", name: "UFPR", label: "Universidade Federal do Paraná" },
+    { index: "101", name: "UFPA", label: "Universidade Federal do Pará" },
+    { index: "102", name: "UFPI", label: "Universidade Federal do Piauí" },
+    {
+      index: "103",
+      name: "UFRB",
+      label: "Universidade Federal do Recôncavo da Bahia",
+    },
+    { index: "104", name: "FURG", label: "Universidade Federal do Rio Grande" },
+    {
+      index: "105",
+      name: "UFRN",
+      label: "Universidade Federal do Rio Grande do Norte",
+    },
+    {
+      index: "106",
+      name: "UFRGS",
+      label: "Universidade Federal do Rio Grande do Sul",
+    },
+    {
+      index: "107",
+      name: "UFRJ",
+      label: "Universidade Federal do Rio de Janeiro",
+    },
+    { index: "108", name: "UFSB", label: "Universidade Federal do Sul da Bahia" },
+    {
+      index: "109",
+      name: "UNIFESSPA",
+      label: "Universidade Federal do Sul e Sudeste do Pará",
+    },
+    {
+      index: "110",
+      name: "UFTM",
+      label: "Universidade Federal do Triângulo Mineiro",
+    },
+    {
+      index: "111",
+      name: "UFVJM",
+      label: "Universidade Federal dos Vales do Jequitinhonha e Mucuri",
+    },
+    { index: "112", name: "FEEVALE", label: "Universidade Feevale" },
+    { index: "113", name: "FUMEC", label: "Universidade Fumec" },
+    { index: "114", name: "UNIB", label: "Universidade Ibirapuera" },
+    { index: "115", name: "UNIG", label: "Universidade Iguaçu" },
+    {
+      index: "116",
+      name: "UNIFENAS",
+      label: "Universidade José do Rosário Vellano",
+    },
+    { index: "117", name: "UNILASALLE", label: "Universidade La Salle" },
+    { index: "118", name: "ULBRA", label: "Universidade Luterana do Brasil" },
+    {
+      index: "119",
+      name: "UNIMEP",
+      label: "Universidade Metodista de Piracicaba",
+    },
+    { index: "120", name: "UMESP", label: "Universidade Metodista de São Paulo" },
+    {
+      index: "121",
+      name: "UNIMES",
+      label: "Universidade Metropolitana de Santos",
+    },
+    {
+      index: "122",
+      name: "USCS",
+      label: "Universidade Municipal de São Caetano do Sul",
+    },
+    { index: "123", name: "UNINILTONLINS", label: "Universidade Nilton Lins" },
+    { index: "124", name: "UNINOVE", label: "Universidade Nove de Julho" },
+    { index: "125", name: "UNIPAR", label: "Universidade Paranaense" },
+    { index: "126", name: "UNIP", label: "Universidade Paulista" },
+    { index: "127", name: "UNOPAR", label: "Universidade Pitágoras Unopar" },
+    { index: "128", name: "UP", label: "Universidade Positivo" },
+    { index: "129", name: "UNP", label: "Universidade Potiguar" },
+    {
+      index: "130",
+      name: "MACKENZIE",
+      label: "Universidade Presbiteriana Mackenzie",
+    },
+    {
+      index: "131",
+      name: "UNIPAC",
+      label: "Universidade Presidente Antônio Carlos",
+    },
+    {
+      index: "132",
+      name: "URI",
+      label: "Universidade Regional Integrada do Alto Uruguai e das Missões",
+    },
+    { index: "133", name: "FURB", label: "Universidade Regional de Blumenau" },
+    { index: "134", name: "URCA", label: "Universidade Regional do Cariri" },
+    {
+      index: "135",
+      name: "UNIJUÍ",
+      label: "Universidade Regional do Noroeste do Estado do Rio Grande do Sul",
+    },
+    { index: "136", name: "UNIVERSO", label: "Universidade Salgado de Oliveira" },
+    { index: "137", name: "UNIFACS", label: "Universidade Salvador" },
+    { index: "138", name: "UNISANTA", label: "Universidade Santa Cecília" },
+    { index: "139", name: "USU", label: "Universidade Santa Úrsula" },
+    { index: "140", name: "UNISA", label: "Universidade Santo Amaro" },
+    { index: "141", name: "USF", label: "Universidade São Francisco" },
+    { index: "142", name: "USJT", label: "Universidade São Judas Tadeu" },
+    {
+      index: "143",
+      name: "UTFPR",
+      label: "Universidade Tecnológica Federal do Paraná",
+    },
+    { index: "144", name: "UNIT", label: "Universidade Tiradentes" },
+    { index: "145", name: "UTP", label: "Universidade Tuiuti do Paraná" },
+    {
+      index: "146",
+      name: "UNG",
+      label: "Universidade Universus Veritas Guarulhos",
+    },
+    { index: "147", name: "UNIVALE", label: "Universidade Vale do Rio Doce" },
+    { index: "148", name: "UNINCOR", label: "Universidade Vale do Rio Verde" },
+    { index: "149", name: "USS", label: "Universidade Vassouras" },
+    { index: "150", name: "UVA RJ", label: "Universidade Veiga de Almeida" },
+    { index: "151", name: "UVV", label: "Universidade Vila Velha" },
+    { index: "152", name: "UNAMA", label: "Universidade da Amazônia" },
+    {
+      index: "153",
+      name: "UNILAB",
+      label:
+        "Universidade da Integração Internacional da Lusofonia Afro-Brasileira",
+    },
+    { index: "154", name: "URCAMP", label: "Universidade da Região da Campanha" },
+    {
+      index: "155",
+      name: "UNIVILLE",
+      label: "Universidade da Região de Joinville",
+    },
+    { index: "156", name: "UNIARA", label: "Universidade de Araraquara" },
+    { index: "157", name: "UNB", label: "Universidade de Brasília" },
+    { index: "158", name: "UCS", label: "Universidade de Caxias do Sul" },
+    { index: "159", name: "UNICRUZ", label: "Universidade de Cruz Alta" },
+    { index: "160", name: "UNIC/UNIME", label: "Universidade de Cuiabá" },
+    { index: "161", name: "UNIFOR CE", label: "Universidade de Fortaleza" },
+    { index: "162", name: "UNIFRAN", label: "Universidade de Franca" },
+    { index: "163", name: "UI", label: "Universidade de Itaúna" },
+    { index: "164", name: "UNIMAR", label: "Universidade de Marília" },
+    { index: "165", name: "UMC", label: "Universidade de Mogi das Cruzes" },
+    { index: "166", name: "UPF", label: "Universidade de Passo Fundo" },
+    { index: "167", name: "UPE", label: "Universidade de Pernambuco" },
+    { index: "168", name: "UNAERP", label: "Universidade de Ribeirão Preto" },
+    { index: "169", name: "FESURV", label: "Universidade de Rio Verde" },
+    { index: "170", name: "UNISC", label: "Universidade de Santa Cruz do Sul" },
+    { index: "171", name: "UNISO", label: "Universidade de Sorocaba" },
+    { index: "172", name: "USP", label: "Universidade de São Paulo" },
+    { index: "173", name: "UNITAU", label: "Universidade de Taubaté" },
+    { index: "174", name: "UNIUBE", label: "Universidade de Uberaba" },
+    { index: "175", name: "UNICEUMA", label: "Universidade do Ceuma" },
+    { index: "176", name: "UNC", label: "Universidade do Contestado" },
+    { index: "177", name: "UNEB", label: "Universidade do Estado da Bahia" },
+    {
+      index: "178",
+      name: "UNEMAT",
+      label: "Universidade do Estado de Mato Grosso",
+    },
+    {
+      index: "179",
+      name: "UEMG",
+      label: "Universidade do Estado de Minas Gerais",
+    },
+    { index: "180", name: "UEAP", label: "Universidade do Estado do Amapá" },
+    { index: "181", name: "UEA", label: "Universidade do Estado do Amazonas" },
+    { index: "182", name: "UEPA", label: "Universidade do Estado do Pará" },
+    {
+      index: "183",
+      name: "UERN",
+      label: "Universidade do Estado do Rio Grande do Norte",
+    },
+    {
+      index: "184",
+      name: "UERJ",
+      label: "Universidade do Estado do Rio de Janeiro",
+    },
+    {
+      index: "185",
+      name: "UNESC SC",
+      label: "Universidade do Extremo Sul Catarinense",
+    },
+    {
+      index: "186",
+      name: "UNIGRANRIO",
+      label: "Universidade do Grande Rio Professor José de Souza Herdy",
+    },
+    { index: "187", name: "UNOESTE", label: "Universidade do Oeste Paulista" },
+    {
+      index: "188",
+      name: "UNOESC",
+      label: "Universidade do Oeste de Santa Catarina",
+    },
+    {
+      index: "189",
+      name: "UNIPLAC",
+      label: "Universidade do Planalto Catarinense",
+    },
+    { index: "190", name: "USC", label: "Universidade do Sagrado Coração" },
+    {
+      index: "191",
+      name: "UNISUL",
+      label: "Universidade do Sul de Santa Catarina",
+    },
+    { index: "192", name: "UNITINS", label: "Universidade do Tocantins" },
+    { index: "193", name: "UNIVALI", label: "Universidade do Vale do Itajaí" },
+    { index: "194", name: "UNIVAP", label: "Universidade do Vale do Paraíba" },
+    {
+      index: "195",
+      name: "UNISINOS",
+      label: "Universidade do Vale do Rio dos Sinos",
+    },
+    { index: "196", name: "UNIVÁS", label: "Universidade do Vale do Sapucaí" },
+    { index: "197", name: "UNIVATES", label: "Universidade do Vale do Taquari" },
+  ];
+
