+++ components/datasets/graduationCourseData.tsx
@@ -1,0 +1,273 @@
+export const graduationCourseData = [
+    {
+      index: "1",
+      name: "Administração",
+    },
+    {
+      index: "2",
+      name: "Arquitetura e Urbanismo",
+    },
+    {
+      index: "3",
+      name: "Artes Cênicas",
+    },
+    {
+      index: "4",
+      name: "Astronomia",
+    },
+    {
+      index: "5",
+      name: "Biblioteconomia e Gestão de Unidades de Informação (CBG)",
+    },
+    {
+      index: "6",
+      name: "Ciências Atuariais ou Estatísticas",
+    },
+    {
+      index: "7",
+      name: "Ciências Biológicas",
+    },
+    {
+      index: "8",
+      name: "Ciências Biológicas Modalidade Biofísica",
+    },
+    {
+      index: "9",
+      name: "Ciências Biológicas Modalidade Médica",
+    },
+    {
+      index: "10",
+      name: "Ciências Biológicas Modalidade Microbiologia e Imunologia",
+    },
+    {
+      index: "11",
+      name: "Ciências Contábeis",
+    },
+    {
+      index: "12",
+      name: "Ciências da Computação",
+    },
+    {
+      index: "13",
+      name: "Ciências Econômicas",
+    },
+    {
+      index: "14",
+      name: "Ciências Sociais",
+    },
+    {
+      index: "15",
+      name: "Composição de Interior",
+    },
+    {
+      index: "16",
+      name: "Composição Paisagística",
+    },
+    {
+      index: "17",
+      name: "Comunicação Social",
+    },
+    {
+      index: "18",
+      name: "Dança",
+    },
+    {
+      index: "19",
+      name: "Desenho Industrial",
+    },
+    {
+      index: "20",
+      name: "Direito",
+    },
+    {
+      index: "21",
+      name: "Direção Teatral",
+    },
+    {
+      index: "22",
+      name: "Educação Física",
+    },
+    {
+      index: "23",
+      name: "Enfermagem e Obstetrícia",
+    },
+    {
+      index: "24",
+      name: "Engenharia Ambiental",
+    },
+    {
+      index: "25",
+      name: "Engenharia Civil",
+    },
+    {
+      index: "26",
+      name: "Engenharia de Alimentos",
+    },
+    {
+      index: "27",
+      name: "Engenharia de Bioprocessos",
+    },
+    {
+      index: "28",
+      name: "Engenharia de Computação e Informação",
+    },
+    {
+      index: "29",
+      name: "Engenharia de Controle e Automação",
+    },
+    {
+      index: "30",
+      name: "Engenharia de Materiais",
+    },
+    {
+      index: "31",
+      name: "Engenharia de Petróleo",
+    },
+    {
+      index: "32",
+      name: "Engenharia de Produção",
+    },
+    {
+      index: "33",
+      name: "Engenharia Elétrica",
+    },
+    {
+      index: "34",
+      name: "Engenharia Eletrônica e de Computação",
+    },
+    {
+      index: "35",
+      name: "Engenharia Mecânica",
+    },
+    {
+      index: "36",
+      name: "Engenharia Metalúrgica",
+    },
+    {
+      index: "37",
+      name: "Engenharia Naval e Oceânica",
+    },
+    {
+      index: "38",
+      name: "Engenharia Química",
+    },
+    {
+      index: "39",
+      name: "Farmácia",
+    },
+    {
+      index: "40",
+      name: "Farmácia - Macaé",
+    },
+    {
+      index: "41",
+      name: "Filosofia",
+    },
+    {
+      index: "42",
+      name: "Física",
+    },
+    {
+      index: "43",
+      name: "Fisioterapia",
+    },
+    {
+      index: "44",
+      name: "Fonoaudiologia",
+    },
+    {
+      index: "45",
+      name: "Geografia",
+    },
+    {
+      index: "46",
+      name: "Geologia",
+    },
+    {
+      index: "47",
+      name: "Gravura",
+    },
+    {
+      index: "48",
+      name: "História",
+    },
+    {
+      index: "49",
+      name: "Letras",
+    },
+    {
+      index: "50",
+      name: "Licenc. em Ciências Biológicas - Macaé",
+    },
+    {
+      index: "51",
+      name: "Licenc. em Educação Artística: Artes Plásticas e Desenho",
+    },
+    {
+      index: "52",
+      name: "Licenciatura em Química - Macaé",
+    },
+    {
+      index: "53",
+      name: "Matemática (Bacharelado e Licenciatura)",
+    },
+    {
+      index: "54",
+      name: "Matemática Aplicada",
+    },
+    {
+      index: "55",
+      name: "Medicina",
+    },
+    {
+      index: "56",
+      name: "Meteorologia",
+    },
+    {
+      index: "57",
+      name: "Microbiologia e Imunologia",
+    },
+    {
+      index: "58",
+      name: "Música",
+    },
+    {
+      index: "59",
+      name: "Nutrição",
+    },
+    {
+      index: "60",
+      name: "Odontologia",
+    },
+    {
+      index: "61",
+      name: "Pedagogia",
+    },
+    {
+      index: "62",
+      name: "Pintura",
+    },
+    {
+      index: "63",
+      name: "Psicologia",
+    },
+    {
+      index: "64",
+      name: "Química",
+    },
+    {
+      index: "65",
+      name: "Química Industrial",
+    },
+    {
+      index: "66",
+      name: "Saúde Coletiva",
+    },
+    {
+      index: "67",
+      name: "Serviço Social",
+    },
+  ];
+  
+  
+
